refactor: Name, Position을 vo 패키지로 이동
