+++ src/main/java/racingcar/vo/Name.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.vo;
 
 import racingcar.util.Validator;
