+++ src/main/java/racingcar/vo/Position.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.vo;
 
 public class Position {
     private static final int FORWARD_ONE_BLOCK = 1;
