+++ src/test/java/racingcar/vo/NameTest.java
@@ -1,10 +1,11 @@
-package racingcar.domain;
+package racingcar.vo;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.vo.Name;
 
 class NameTest {
