+++ src/test/java/racingcar/vo/PositionTest.java
@@ -1,9 +1,10 @@
-package racingcar.domain;
+package racingcar.vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.vo.Position;
 
 class PositionTest {
